Move GriddedPSFModel to new module
