+++ photutils/psf/models.py
@@ -4,20 +4,17 @@ This module provides models for doing PSF/PRF-fitting photometry.
 """
 
 import copy
-import itertools
 import warnings
-from functools import lru_cache
 
 import numpy as np
 from astropy.modeling import Fittable2DModel, Parameter
-from astropy.nddata import NDData
 from astropy.utils.exceptions import AstropyWarning
 
 from photutils.aperture import CircularAperture
 from photutils.utils._parameters import as_pair
 
 __all__ = ['NonNormalizable', 'FittableImageModel', 'EPSFModel',
-           'GriddedPSFModel', 'IntegratedGaussianPRF', 'PRFAdapter']
+           'IntegratedGaussianPRF', 'PRFAdapter']
 
 
 class NonNormalizable(AstropyWarning):
@@ -678,303 +675,6 @@ class EPSFModel(FittableImageModel):
                        | ((yi < 0) | (yi > (self._ny - 1)
                                       / self.oversampling[0])))
             evaluated_model[invalid] = self._fill_value
-
-        return evaluated_model
-
-
-class GriddedPSFModel(Fittable2DModel):
-    """
-    A fittable 2D model containing a grid PSF models defined at specific
-    locations that are interpolated to evaluate a PSF at an arbitrary
-    (x, y) position.
-
-    Parameters
-    ----------
-    data : `~astropy.nddata.NDData`
-        A `~astropy.nddata.NDData` object containing the grid of
-        reference PSF arrays. The data attribute must contain a 3D
-        `~numpy.ndarray` containing a stack of the 2D PSFs with a shape
-        of ``(N_psf, PSF_ny, PSF_nx)``. The meta attribute must be
-        `dict` containing the following:
-
-            * ``'grid_xypos'``: A list of the (x, y) grid positions of
-              each reference PSF. The order of positions should match the
-              first axis of the 3D `~numpy.ndarray` of PSFs. In other
-              words, ``grid_xypos[i]`` should be the (x, y) position of
-              the reference PSF defined in ``data[i]``.
-
-            * ``'oversampling'``: The integer oversampling factor of the
-              PSF.
-
-        The meta attribute may contain other properties such as the
-        telescope, instrument, detector, and filter of the PSF.
-    """
-
-    flux = Parameter(description='Intensity scaling factor for the PSF '
-                     'model.', default=1.0)
-    x_0 = Parameter(description='x position in the output coordinate grid '
-                    'where the model is evaluated.', default=0.0)
-    y_0 = Parameter(description='y position in the output coordinate grid '
-                    'where the model is evaluated.', default=0.0)
-
-    def __init__(self, data, *, flux=flux.default, x_0=x_0.default,
-                 y_0=y_0.default, fill_value=0.0):
-
-        self._data_input = self._validate_data(data)
-        self.data = data.data
-        self._meta = data.meta  # use _meta to avoid the meta descriptor
-        self.grid_xypos = data.meta['grid_xypos']
-        self.oversampling = data.meta['oversampling']
-        self.fill_value = fill_value
-
-        self._grid_xpos, self._grid_ypos = np.transpose(self.grid_xypos)
-        self._xgrid = np.unique(self._grid_xpos)  # also sorts values
-        self._ygrid = np.unique(self._grid_ypos)  # also sorts values
-        if (len(list(itertools.product(self._xgrid, self._ygrid)))
-                != len(self.grid_xypos)):
-            raise ValueError('"grid_xypos" must form a regular grid.')
-
-        self._xidx = np.arange(self.data.shape[2], dtype=float)
-        self._yidx = np.arange(self.data.shape[1], dtype=float)
-
-        # Here we avoid decorating the instance method with @lru_cache
-        # to prevent memory leaks; we set maxsize=128 to prevent the
-        # cache from growing too large.
-        self._calc_interpolator = lru_cache(maxsize=128)(
-            self._calc_interpolator_uncached)
-
-        super().__init__(flux, x_0, y_0)
-
-    @staticmethod
-    def _validate_data(data):
-        if not isinstance(data, NDData):
-            raise TypeError('data must be an NDData instance.')
-
-        if data.data.ndim != 3:
-            raise ValueError('The NDData data attribute must be a 3D numpy '
-                             'ndarray')
-
-        if 'grid_xypos' not in data.meta:
-            raise ValueError('"grid_xypos" must be in the nddata meta '
-                             'dictionary.')
-        if len(data.meta['grid_xypos']) != data.data.shape[0]:
-            raise ValueError('The length of grid_xypos must match the number '
-                             'of input PSFs.')
-
-        if 'oversampling' not in data.meta:
-            raise ValueError('"oversampling" must be in the nddata meta '
-                             'dictionary.')
-        if not np.isscalar(data.meta['oversampling']):
-            raise ValueError('oversampling must be a scalar value')
-
-        return data
-
-    def copy(self):
-        """
-        Return a copy of this model.
-
-        Note that the PSF grid data is not copied. Use the `deepcopy`
-        method if you want to copy the PSF grid data.
-        """
-        return self.__class__(self._data_input, flux=self.flux.value,
-                              x_0=self.x_0.value, y_0=self.y_0.value,
-                              fill_value=self.fill_value)
-
-    def deepcopy(self):
-        """
-        Return a deep copy of this model.
-        """
-        return copy.deepcopy(self)
-
-    def clear_cache(self):
-        """
-        Clear the internal cache.
-        """
-        self._calc_interpolator.cache_clear()
-
-    def _cache_info(self):
-        """
-        Return information about the internal cache.
-        """
-        return self._calc_interpolator.cache_info()
-
-    @staticmethod
-    def _find_start_idx(data, x):
-        """
-        Find the index of the lower bound where ``x`` should be inserted
-        into ``a`` to maintain order.
-
-        The index of the upper bound is the index of the lower bound
-        plus 2.  Both bound indices must be within the array.
-
-        Parameters
-        ----------
-        data : 1D `~numpy.ndarray`
-            The 1D array to search.
-
-        x : float
-            The value to insert.
-
-        Returns
-        -------
-        index : int
-            The index of the lower bound.
-        """
-        idx = np.searchsorted(data, x)
-        if idx == 0:
-            idx0 = 0
-        elif idx == len(data):  # pragma: no cover
-            idx0 = idx - 2
-        else:
-            idx0 = idx - 1
-        return idx0
-
-    def _find_bounding_points(self, x, y):
-        """
-        Find the indices of the grid points that bound the input
-        ``(x, y)`` position.
-
-        Parameters
-        ----------
-        x, y : float
-            The ``(x, y)`` position where the PSF is to be evaluated.
-            The position must be inside the region defined by the grid
-            of PSF positions.
-
-        Returns
-        -------
-        indices : list of int
-            A list of indices of the bounding grid points.
-        """
-        x0 = self._find_start_idx(self._xgrid, x)
-        y0 = self._find_start_idx(self._ygrid, y)
-        xypoints = list(itertools.product(self._xgrid[x0:x0 + 2],
-                                          self._ygrid[y0:y0 + 2]))
-
-        # find the grid_xypos indices of the reference xypoints
-        indices = []
-        for xx, yy in xypoints:
-            indices.append(np.argsort(np.hypot(self._grid_xpos - xx,
-                                               self._grid_ypos - yy))[0])
-
-        return indices
-
-    @staticmethod
-    def _bilinear_interp(xyref, zref, xi, yi):
-        """
-        Perform bilinear interpolation of four 2D arrays located at
-        points on a regular grid.
-
-        Parameters
-        ----------
-        xyref : list of 4 (x, y) pairs
-            A list of 4 ``(x, y)`` pairs that form a rectangle.
-
-        zref : 3D `~numpy.ndarray`
-            A 3D `~numpy.ndarray` of shape ``(4, nx, ny)``. The first
-            axis corresponds to ``xyref``, i.e., ``refdata[0, :, :]`` is
-            the 2D array located at ``xyref[0]``.
-
-        xi, yi : float
-            The ``(xi, yi)`` point at which to perform the
-            interpolation.  The ``(xi, yi)`` point must lie within the
-            rectangle defined by ``xyref``.
-
-        Returns
-        -------
-        result : 2D `~numpy.ndarray`
-            The 2D interpolated array.
-        """
-        xyref = [tuple(i) for i in xyref]
-        idx = sorted(range(len(xyref)), key=xyref.__getitem__)
-        xyref = sorted(xyref)  # sort by x, then y
-        (x0, y0), (_x0, y1), (x1, _y0), (_x1, _y1) = xyref
-
-        if x0 != _x0 or x1 != _x1 or y0 != _y0 or y1 != _y1:
-            raise ValueError('The refxy points do not form a rectangle.')
-
-        if not np.isscalar(xi):
-            xi = xi[0]
-        if not np.isscalar(yi):
-            yi = yi[0]
-
-        if not x0 <= xi <= x1 or not y0 <= yi <= y1:
-            raise ValueError('The (x, y) input is not within the rectangle '
-                             'defined by xyref.')
-
-        data = np.asarray(zref)[idx]
-        weights = np.array([(x1 - xi) * (y1 - yi), (x1 - xi) * (yi - y0),
-                            (xi - x0) * (y1 - yi), (xi - x0) * (yi - y0)])
-        norm = (x1 - x0) * (y1 - y0)
-
-        return np.sum(data * weights[:, None, None], axis=0) / norm
-
-    def _calc_interpolator_uncached(self, x_0, y_0):
-        """
-        Return the local interpolation function for the PSF model at
-        (x_0, y_0).
-
-        Note that the interpolator will be cached by _calc_interpolator.
-        It can be cleared by calling the clear_cache method.
-        """
-        from scipy.interpolate import RectBivariateSpline
-
-        if (x_0 < self._xgrid[0] or x_0 > self._xgrid[-1]
-                or y_0 < self._ygrid[0] or y_0 > self._ygrid[-1]):
-            # position is outside of the grid, so simply use the
-            # closest reference PSF
-            ref_index = np.argsort(np.hypot(self._grid_xpos - x_0,
-                                            self._grid_ypos - y_0))[0]
-            psf_image = self.data[ref_index, :, :]
-        else:
-            # find the four bounding reference PSFs and interpolate
-            ref_indices = self._find_bounding_points(x_0, y_0)
-            xyref = np.array(self.grid_xypos)[ref_indices]
-            psfs = self.data[ref_indices, :, :]
-
-            psf_image = self._bilinear_interp(xyref, psfs, x_0, y_0)
-
-        interpolator = RectBivariateSpline(self._xidx, self._yidx,
-                                           psf_image.T, kx=3, ky=3, s=0)
-
-        return interpolator
-
-    def evaluate(self, x, y, flux, x_0, y_0):
-        """
-        Evaluate the `GriddedPSFModel` for the input parameters.
-        """
-        # NOTE: the astropy base Model.__call__() method converts scalar
-        # inputs to size-1 arrays before calling evaluate().
-        if not np.isscalar(flux):
-            flux = flux[0]
-        if not np.isscalar(x_0):
-            x_0 = x_0[0]
-        if not np.isscalar(y_0):
-            y_0 = y_0[0]
-
-        # Calculate the local interpolation function for the PSF at
-        # (x_0, y_0). Only the integer part of the position is input in
-        # order to have effective caching.
-        interpolator = self._calc_interpolator(int(x_0), int(y_0))
-
-        # now evaluate the PSF at the (x_0, y_0) subpixel position on
-        # the input (x, y) values
-        xi = self.oversampling * (np.asarray(x, dtype=float) - x_0)
-        yi = self.oversampling * (np.asarray(y, dtype=float) - y_0)
-
-        # define origin at the PSF image center
-        ny, nx = self.data.shape[1:]
-        xi += (nx - 1) / 2
-        yi += (ny - 1) / 2
-
-        evaluated_model = flux * interpolator.ev(xi, yi)
-
-        if self.fill_value is not None:
-            # find indices of pixels that are outside the input pixel
-            # grid and set these pixels to the fill_value
-            invalid = (((xi < 0) | (xi > nx - 1))
-                       | ((yi < 0) | (yi > ny - 1)))
-            evaluated_model[invalid] = self.fill_value
 
         return evaluated_model
