+++ photutils/psf/tests/test_models.py
@@ -11,8 +11,8 @@ from astropy.modeling.models import Gaussian2D, Moffat2D
 from astropy.nddata import NDData
 from numpy.testing import assert_allclose
 
-from photutils.psf.models import (FittableImageModel, GriddedPSFModel,
-                                  IntegratedGaussianPRF, PRFAdapter)
+from photutils.psf import (FittableImageModel, GriddedPSFModel,
+                           IntegratedGaussianPRF, PRFAdapter)
 from photutils.segmentation import SourceCatalog, detect_sources
 from photutils.utils._optional_deps import HAS_SCIPY
